Version 1.3.1 (Build 12)  Reading purchase receipt  Truncating middle in the table cells

+++ File Catalyst/Security/Receipts.h
@@ -9,6 +9,8 @@
 #ifndef Caravelle_Receipts_h
 #define Caravelle_Receipts_h
 
-NSInteger validateAppReceipt ();
+
+int validateAppReceipt ();
+
 
 #endif
